break out scipy optimization classes
