+++ floris/tools/optimization/scipy/__init__.py
@@ -1,1 +1,9 @@
+from . import base_COE
+from . import layout
+from . import layout_height
 from . import optimization
+from . import power_density
+from . import power_density_1D
+from . import yaw
+from . import yaw_wind_rose
+from . import yaw_wind_rose_parallel
